Fixed random tree training algorithm
The training is now performed with respect to random forest training. It
means that five random node split funtion parameters are generated (e.g.
function type, threshold, channel) and for every of these function
proposals, the information gain is computed. Then, the function with
highes information gain is chosen to be used.

Resolves #3